consolidate math package
