--- pydl/pydlutils/math/__init__.py
@@ -1,9 +1,0 @@
-# Licensed under a 3-clause BSD style license - see LICENSE.rst
-# -*- coding: utf-8 -*-
-"""
-This subpackage corresponds to the math directory of idlutils.
-"""
-from .computechi2 import computechi2
-from .djs_median import djs_median
-from .djs_reject import djs_reject
-from .find_contiguous import find_contiguous
