+++ pydl/pydlutils/tests/test_math.py
@@ -1,16 +1,52 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 # -*- coding: utf-8 -*-
+import numpy as np
+from astropy.tests.helper import raises
+from ..math import computechi2, djs_median
 
-def test_djs_median():
-    from numpy import allclose, array, median
-    from numpy.random import random, seed
-    from astropy.tests.helper import raises
-    from .. import djs_median
-    seed(424242)
-    data = 100.0*random(100)
-    data2 = 100.0*random((10, 10))
-    data3 = 100.0*random((10, 10, 10))
-    data_width_5 = array([
+
+class TestMath(object):
+    """Test the functions in pydl.pydlutils.math.
+    """
+
+    def setup(self):
+        pass
+
+    def teardown(self):
+        pass
+
+    def test_computechi2(self):
+        x = np.arange(20)
+        y = np.array([6.6198438, 1.3491303, 0.41035045, 9.4013375, 4.1103360,
+            4.3522868, 4.6338078, 4.7400367, 5.1860726, 5.1082748,
+            5.1084596, 5.2990997, 5.5987537, 5.7007504, 5.7855296,
+            4.1123709, 6.9437957, 4.9956179, 4.3724215, 3.6245063])
+        sqivar = np.array([1.32531, 0.886090, 1.08384, 1.04489, 1.46807,
+            1.30800, 0.507725, 1.12840, 0.955025, 1.35925,
+            1.10126, 1.45690, 0.575700, 0.949710, 1.23368,
+            0.536489, 0.772543, 0.957729, 0.883976, 1.11559])
+        templates = np.vstack((np.ones((20,),dtype='d'),x)).transpose()
+        chi2 = computechi2(y,sqivar,templates)
+        #
+        # 20 data points, 2 parameters = 18 degrees of freedom.
+        #
+        assert chi2.dof == 18
+        #
+        # The covariance matrix should be symmetric
+        #
+        assert (chi2.covar.T == chi2.covar).all()
+        #
+        # The variance vector should be the same as the diagonal of the
+        # covariance matrix.
+        #
+        assert (chi2.var == np.diag(chi2.covar)).all()
+
+def test_djs_median(self):
+    np.random.seed(424242)
+    data = 100.0*np.random.random(100)
+    data2 = 100.0*np.random.random((10, 10))
+    data3 = 100.0*np.random.random((10, 10, 10))
+    data_width_5 = np.array([
         49.68311576,  74.83671757,  49.68311576,  42.12573137,
         32.11576752,  27.22092569,  15.08905903,  15.08905903,
         27.22092569,  20.07809411,  41.55978953,  41.55978953,
@@ -36,7 +72,7 @@ def test_djs_median():
         34.55273356,  62.09010468,  62.09010468,  70.43134908,
         68.89705132,  68.89705132,  68.89705132,  66.30426084,
         55.92748086,  55.92748086,  55.92748086,  65.11387444])
-    data_width_5_reflect = array([
+    data_width_5_reflect = np.array([
         49.68311576,  49.68311576,  49.68311576,  42.12573137,
         32.11576752,  27.22092569,  15.08905903,  15.08905903,
         27.22092569,  20.07809411,  41.55978953,  41.55978953,
@@ -65,15 +101,17 @@ def test_djs_median():
     #
     # Degenerate cases that fall back on numpy.median().
     #
-    assert allclose(median(data), djs_median(data))
-    assert allclose(median(data2, axis=0), djs_median(data2, dimension=0))
-    assert allclose(median(data2, axis=1), djs_median(data2, dimension=1))
+    assert np.allclose(np.median(data), djs_median(data))
+    assert np.allclose(np.median(data2, axis=0),
+                        djs_median(data2, dimension=0))
+    assert np.allclose(np.median(data2, axis=1),
+                        djs_median(data2, dimension=1))
     #
     # Test widths.
     #
-    assert allclose(data, djs_median(data, width=1))
-    assert allclose(data_width_5, djs_median(data, width=5))
-    # assert allclose(data_width_5_reflect,
+    assert np.allclose(data, djs_median(data, width=1))
+    assert np.allclose(data_width_5, djs_median(data, width=5))
+    # assert np.allclose(data_width_5_reflect,
     #                 djs_median(data, width=5, boundary='reflect'))
     #
     # Exceptions
